requirements.txt
Flask
flask-cors
google-cloud-speech
